% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprac-package.R
\docType{package}
\name{emprac-package}
\alias{emprac}
\alias{emprac-package}
\title{emprac: Demo R Package}
\description{
A demo of how to make a simple R package.

This package is an exercise in package creation using R studio. The package includes two sample datasets with their respective documentation.
}
\references{
\url{https://github.com/phille10/emprac}
}
\author{
\strong{Maintainer}: Elizabeth Phillips \email{phille10@mcmaster.ca}


Elizabeth Phillips \email{phille10@mcmaster.ca}
}
\keyword{internal}
