% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprac-package.R
\docType{data}
\name{pitchclass_inacc}
\alias{pitchclass_inacc}
\title{Pitch class inaccuracy}
\format{
A data frame with 2047 rows and 7 columns:
\describe{
\item{X}{Index of song per region}
\item{Unnamed..0}{Index of analyzed pitch class per song}
\item{Unnamed..0.1}{Names of analyzed pitch class per song}
\item{song.id}{Subregion/Name of song}
\item{instrument.category}{Type of instrument (Vocal)}
\item{region.category}{Region of song (from 10 global regions)}
\item{pitch.class.inaccuracy}{Inaccuracy of pitch class in cents displacement from nearest 100-cent multiple of the tonic}
}
}
\source{
Vocal imprecision as a universal constraint on the structure of musical scales, data archive (https://osf.io/jn78c/) Accessed: November 1, 2023
data(pitchclass_inacc)
inaccuracy <- pitchclass_inacc$pitch.class.inaccuracy
}
\usage{
data(pitchclass_inacc)
}
\description{
A dataset containing information about the pitch class inaccuracy (relative to 12-ET) of 418 songs worldwide
}
\keyword{datasets}
