% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emprac-package.R
\docType{data}
\name{pitchclass_range}
\alias{pitchclass_range}
\title{Pitch class imprecision}
\format{
A data frame with 2458 rows and 7 columns:
\describe{
\item{X}{Index of song per region}
\item{Unnamed..0}{Index of analyzed pitch class per song}
\item{Unnamed..0.1}{Names of analyzed pitch class per song}
\item{song.id}{Subregion/Name of song}
\item{instrument.category}{Type of instrument (Vocal)}
\item{region.category}{Region of song (from 10 global regions)}
\item{pitch.class.range}{Imprecision (range) of pitch class in cents}
}
}
\source{
Vocal imprecision as a universal constraint on the structure of musical scales, data archive (https://osf.io/jn78c/) Accessed: November 1, 2023
data(pitchclass_range)
imprecision <- pitchclass_range$pitch.class.imprecision
}
\usage{
data(pitchclass_range)
}
\description{
A dataset containing information about the pitch class imprecision of 418 songs worldwide
}
\keyword{datasets}
